build.xml: include *.jar and *.jar.pack.gz in deployment
java has a utility called pack200 which is a compression utility
that takes an existing compressed jar and further compresses it down
by a nice factor.  with the new version of oracle's java for the mac,
this packed jar fails to be read by the client. it's unfortunate,
but it means we must keep both the non packed and packed jar files.