felix.jar: new oracle java for the mac exposes buggy logic in native loader.
this is another unfortunate side effect of the new version of java that
oracle recently published for the mac. it returned a system property
value that didn't jive very well with the existing felix logic.
this means we need to carry around a slightly modified version of felix 3.2.2

tagged: release-1.0 in the web-felix repo